fix: second step
