+++ src/api/axiosConfig/getPortalStaffQuery/index.ts
@@ -4,8 +4,6 @@ import {
   fetchTimeoutServices,
   maxRetriesServices,
 } from "@config/environment";
-import { IBusinessUnitsPortalStaff } from "@ptypes/staffPortalBusiness.types";
-import { IBusinessUnitsPortalStaffId } from "@ptypes/staffBusinessManagersId";
 
 const axiosInstance: AxiosInstance = axios.create({
   baseURL: enviroment.IVITE_IPORTAL_STAFF_QUERY_PROCESS_SERVICE,
@@ -52,36 +50,4 @@ const fetchWithRetries = async <T>(
   throw new Error("Error al obtener los de las unidades de negocio.");
 };
 
-const getBusinessUnitsPortalStaff = async (
-  portalPublicCode: string,
-  userAccount: string
-): Promise<IBusinessUnitsPortalStaff[]> => {
-  const config: AxiosRequestConfig = {
-    headers: {
-      "X-Action": "SearchBusinessUnitsForAnOfficer",
-    },
-  };
-  return fetchWithRetries<IBusinessUnitsPortalStaff[]>(
-    `/business-units-portal-staff/${userAccount}/${portalPublicCode}`,
-    config
-  );
-};
-
-const getBusinessManagersId = async (
-  publiCode: string
-): Promise<IBusinessUnitsPortalStaffId[]> => {
-  const queryParams = new URLSearchParams({
-    publiCode: publiCode,
-  });
-  const config: AxiosRequestConfig = {
-    headers: {
-      "X-Action": "SearchAllBusinessManagerIportalStaff",
-    },
-  };
-  return fetchWithRetries<IBusinessUnitsPortalStaffId[]>(
-    `/staffs?${queryParams.toString()}`,
-    config
-  );
-};
-
-export { getBusinessUnitsPortalStaff, getBusinessManagersId };
+export { axiosInstance, fetchWithRetries };
