+++ src/design/templates/AssignmentForm/interface.tsx
@@ -1,7 +1,7 @@
+import { MdSearch } from "react-icons/md";
 import { Button } from "@inubekit/button";
 import { Stack } from "@inubekit/stack";
 import { Textfield } from "@inubekit/textfield";
-import { MdSearch } from "react-icons/md";
 import {
   Col,
   Colgroup,
@@ -15,10 +15,10 @@ import {
 import { Toggle } from "@inubekit/toggle";
 import { Text } from "@inubekit/text";
 import { Fieldset } from "@inubekit/fieldset";
-import { StyledForm, StyledToggleContainer } from "./styles";
-import { IAssignmentFormUI, titlesOptions } from "./types";
 import { MultipleChoices } from "@design/navigation/MultipleChoices";
 import { basic } from "@design/tokens";
+import { StyledForm, StyledToggleContainer } from "./styles";
+import { IAssignmentFormUI, titlesOptions } from "./types";
 
 const AssignmentFormUI = (props: IAssignmentFormUI) => {
   const {
