+++ src/services/staffPortal/getBusinessManager/index.tsx
@@ -1,5 +1,5 @@
 import { IBusinessManagers } from "@ptypes/staffPortal.types";
-import { getBusinessManagers } from "@api/isaasQuery";
+import { getBusinessManagers } from "@api/getBusinessManagers";
 import { mapBusinessManagerApiToEntity } from "./mappers";
 
 const businessManagers = async (
